Organize all classes into suitable java packages
