+++ data/stelle.txt
@@ -1,2 +1,3 @@
 E | 1 | SoC Career Fair | 2024-08-30 10:00to2024-08-31 17:00
-D | 0 | LAJ | 2024-08-28 23:59
+D | 1 | LAJ | 2024-08-28 23:59
+T | 0 | Read LAJ1201
