+++ src/main/java/stelle/Storage.java
@@ -1,3 +1,13 @@
+package stelle;
+
+import stelle.exception.StorageWriteException;
+import stelle.exception.TaskException;
+import stelle.exception.TaskWithoutTypeException;
+import stelle.task.Deadline;
+import stelle.task.Event;
+import stelle.task.Task;
+import stelle.task.ToDo;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.FileNotFoundException;
@@ -8,7 +18,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 /**
- * Represents a Storage object that handles writing and reading tasks from a local file.
+ * Represents a stelle.Storage object that handles writing and reading tasks from a local file.
  * @author Lee Ze Hao (A0276123J)
  */
 
@@ -16,7 +26,7 @@ public class Storage {
     private final String filePath;
 
     /**
-     * Creates a new Storage object.
+     * Creates a new stelle.Storage object.
      * @param filePath The file path data is stored to and read from.
      */
     public Storage(String filePath) {
@@ -51,13 +61,13 @@ public class Storage {
                 textToWrite = textToWrite + "T | " + isDoneMark + " | " + task.getTaskName() + "\n";
                 break;
             case DEADLINE:
-                // Note: taskType only set to DEADLINE when Deadline class is created, and is final, so this is safe
+                // Note: taskType only set to DEADLINE when stelle.task.Deadline class is created, and is final, so this is safe
                 Deadline deadline = (Deadline) task;
                 textToWrite = textToWrite + "D | " + isDoneMark + " | " + deadline.getTaskName() +
                         " | " + deadline.getByTime() + "\n";
                 break;
             case EVENT:
-                // Note: taskType only set to EVENT when Event class is created, and is final, so this is safe
+                // Note: taskType only set to EVENT when stelle.task.Event class is created, and is final, so this is safe
                 Event event = (Event) task;
                 textToWrite = textToWrite + "E | " + isDoneMark + " | " + event.getTaskName() +
                         " | " + event.getFromTime() + "to" + event.getToTime() + "\n";
@@ -77,7 +87,7 @@ public class Storage {
 
     /**
      * Reads a list of tasks from a text file at the filePath given during object creation.
-     * @return ArrayList<Task> List of tasks stored in the text file.
+     * @return ArrayList<stelle.task.Task> List of tasks stored in the text file.
      */
     public ArrayList<Task> readTasksFromFile() throws FileNotFoundException, TaskWithoutTypeException {
         ArrayList<Task> taskList = new ArrayList<Task>();
