+++ src/main/java/stelle/exception/DeadlineNoDescriptionException.java
@@ -1,4 +1,6 @@
-/** Exception thrown when there is no description given during creation of a deadline. Child of TaskException.
+package stelle.exception;
+
+/** Exception thrown when there is no description given during creation of a deadline. Child of stelle.exception.TaskException.
  * @author Lee Ze Hao (A0276123J)
  */
