+++ src/main/java/stelle/exception/DeletionNotSpecifiedException.java
@@ -1,4 +1,6 @@
-/** Exception thrown when there is no list number given during deletion command. Child of StelleException.
+package stelle.exception;
+
+/** Exception thrown when there is no list number given during deletion command. Child of stelle.exception.StelleException.
  * @author Lee Ze Hao (A0276123J)
  */
