+++ src/main/java/stelle/exception/StelleException.java
@@ -1,3 +1,5 @@
+package stelle.exception;
+
 /** Parent for all exceptions related to the chatbot.
  * @author Lee Ze Hao (A0276123J)
  */
