+++ src/main/java/stelle/exception/TaskException.java
@@ -1,4 +1,6 @@
-/** Parent for all exceptions related to tasks. Child of StelleException.
+package stelle.exception;
+
+/** Parent for all exceptions related to tasks. Child of stelle.exception.StelleException.
  * @author Lee Ze Hao (A0276123J)
  */
