+++ src/main/java/stelle/exception/ToDoNoDescriptionException.java
@@ -1,4 +1,6 @@
-/** Exception thrown when there is no description given during creation of a ToDo. Child of TaskException.
+package stelle.exception;
+
+/** Exception thrown when there is no description given during creation of a stelle.task.ToDo. Child of stelle.exception.TaskException.
  * @author Lee Ze Hao (A0276123J)
  */
