+++ src/main/java/stelle/exception/WrongDateFormatException.java
@@ -1,5 +1,7 @@
+package stelle.exception;
+
 /** Exception thrown when the date given by user for a task is not of the following format.
- * Child of StelleException.
+ * Child of stelle.exception.StelleException.
  * @author Lee Ze Hao (A0276123J)
  */
