+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Stelle.
+ Hello! I'm stelle.
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
